# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Satellite Image Loader","description":"Service to interact with satellite images on a database","version":"1.0"},"host":"localhost:80","basePath":"/image-loader/api/v1","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/processed-images":{"get":{"tags":["Processed images"],"summary":"Get processed satellite image Processed images","description":"get a processed image from the database","operationId":"Processed images#Get processed satellite image","parameters":[{"name":"Get Processed Satellite ImageRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/ProcessedImagesGetProcessedSatelliteImageRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/ProcessedImagesGetProcessedSatelliteImageResponseBody"}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/ProcessedImagesGetProcessedSatelliteImageBadRequestResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/ProcessedImagesGetProcessedSatelliteImageInternalErrorResponseBody"}}},"schemes":["http"]},"post":{"tags":["Processed images"],"summary":"Load new processed satellite image Processed images","description":"loads a new processed image into the database","operationId":"Processed images#Load new processed satellite image","parameters":[{"name":"Load New Processed Satellite ImageRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/ProcessedImagesLoadNewProcessedSatelliteImageRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/ProcessedImagesLoadNewProcessedSatelliteImageResponseBody"}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/ProcessedImagesLoadNewProcessedSatelliteImageBadRequestResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/ProcessedImagesLoadNewProcessedSatelliteImageInternalErrorResponseBody"}}},"schemes":["http"]}},"/raw-images":{"get":{"tags":["Raw images"],"summary":"Get raw satellite image Raw images","description":"get a raw image from the database","operationId":"Raw images#Get raw satellite image","parameters":[{"name":"Get Raw Satellite ImageRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/RawImagesGetRawSatelliteImageRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/RawImagesGetRawSatelliteImageResponseBody"}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/RawImagesGetRawSatelliteImageBadRequestResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/RawImagesGetRawSatelliteImageInternalErrorResponseBody"}}},"schemes":["http"]},"post":{"tags":["Raw images"],"summary":"Load new raw satellite image Raw images","description":"loads a new image into the database","operationId":"Raw images#Load new raw satellite image","parameters":[{"name":"Load New Raw Satellite ImageRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/RawImagesLoadNewRawSatelliteImageRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/RawImagesLoadNewRawSatelliteImageResponseBody"}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/RawImagesLoadNewRawSatelliteImageBadRequestResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/RawImagesLoadNewRawSatelliteImageInternalErrorResponseBody"}}},"schemes":["http"]}}},"definitions":{"GeographicInformationRequestBody":{"title":"GeographicInformationRequestBody","type":"object","properties":{"coordinates":{"type":"object","description":"Coordinates of the satellite image","example":{"Omnis omnis quasi necessitatibus.":0.6163929175677805},"additionalProperties":true},"tag_name":{"type":"string","description":"Non-forgetable identifier","example":"Nemo ut assumenda perferendis est harum."}},"example":{"coordinates":{"Facere qui dolores quia iusto sint.":0.006707504438042929},"tag_name":"Alias modi molestiae dolorem quas numquam."}},"NormalizedIndexesRequestBody":{"title":"NormalizedIndexesRequestBody","type":"object","properties":{"ndvi":{"type":"array","items":{"type":"number","example":0.4376500901841957,"format":"double"},"description":"Normalized difference vegetation index","example":[0.2484626358366304,0.18745658491395728]},"ndwi":{"type":"array","items":{"type":"number","example":0.5592610830362001,"format":"double"},"description":"Normalized difference water index","example":[0.4119254190513769,0.8639196191267979,0.62920999890917]}},"example":{"ndvi":[0.4820876163670806,0.47616026717394694,0.6377983805667723],"ndwi":[0.1828987573968838,0.0401356398891894,0.6971322235548086]}},"ProcessedImagesGetProcessedSatelliteImageBadRequestResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Get processed satellite image_BadRequest_Response_Body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"ProcessedImagesGetProcessedSatelliteImageInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Get processed satellite image_InternalError_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ProcessedImagesGetProcessedSatelliteImageRequestBody":{"title":"ProcessedImagesGetProcessedSatelliteImageRequestBody","type":"object","properties":{"file_name":{"type":"string","description":"File name of the processed image","example":"Enim quia doloremque nulla."}},"example":{"file_name":"Accusantium iusto in reprehenderit quae aut."}},"ProcessedImagesGetProcessedSatelliteImageResponseBody":{"title":"Mediatype identifier: vnd.application/goa.result+json; view=default","type":"object","properties":{"code":{"type":"string","description":"The operation code","example":"Id dolorum."},"description":{"type":"string","description":"The operation description","example":"Qui et nemo sint facilis."}},"description":"Get Processed Satellite ImageResponseBody result type (default view)","example":{"code":"Omnis alias.","description":"Ea quidem sint dolores neque ut corrupti."}},"ProcessedImagesLoadNewProcessedSatelliteImageBadRequestResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Load new processed satellite image_BadRequest_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"ProcessedImagesLoadNewProcessedSatelliteImageInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Load new processed satellite image_InternalError_Response_Body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"ProcessedImagesLoadNewProcessedSatelliteImageRequestBody":{"title":"ProcessedImagesLoadNewProcessedSatelliteImageRequestBody","type":"object","properties":{"date_time":{"type":"string","description":"When was the image taken","example":"1974-03-04T19:57:34Z","format":"date-time"},"file_name":{"type":"string","description":"File name of the processed image","example":"Est corporis autem omnis."},"geographic_information":{"$ref":"#/definitions/GeographicInformationRequestBody"},"id":{"type":"string","description":"The image identifier","example":"Eos qui mollitia explicabo est corporis officia."},"normalized_indexes":{"$ref":"#/definitions/NormalizedIndexesRequestBody"}},"example":{"date_time":"1994-09-11T21:53:24Z","file_name":"Qui qui.","geographic_information":{"coordinates":{"Ex exercitationem aut reiciendis quod voluptate incidunt.":0.7392343776263494,"Molestiae sed molestias praesentium rerum.":0.3908551358201754,"Qui qui iusto praesentium.":0.19181675265671314},"tag_name":"Incidunt odit tempore."},"id":"Assumenda temporibus.","normalized_indexes":{"ndvi":[0.6127802172492277,0.004560762359582058,0.480367361557016],"ndwi":[0.08253636129242008,0.030755381445945546,0.659276994468877]}}},"ProcessedImagesLoadNewProcessedSatelliteImageResponseBody":{"title":"Mediatype identifier: vnd.application/goa.result+json; view=default","type":"object","properties":{"code":{"type":"string","description":"The operation code","example":"Dolorem laborum illum facere autem."},"description":{"type":"string","description":"The operation description","example":"Corrupti eligendi animi id tempora voluptatem recusandae."}},"description":"Load New Processed Satellite ImageResponseBody result type (default view)","example":{"code":"Quod incidunt laboriosam est aliquam ratione impedit.","description":"Magni maxime."}},"RawImagesGetRawSatelliteImageBadRequestResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Get raw satellite image_BadRequest_Response_Body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"RawImagesGetRawSatelliteImageInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Get raw satellite image_InternalError_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"RawImagesGetRawSatelliteImageRequestBody":{"title":"RawImagesGetRawSatelliteImageRequestBody","type":"object","properties":{"file_name":{"type":"string","description":"File name of the raw image","example":"Nostrum at."}},"example":{"file_name":"Voluptatem repellat et."}},"RawImagesGetRawSatelliteImageResponseBody":{"title":"Mediatype identifier: vnd.application/goa.result+json; view=default","type":"object","properties":{"code":{"type":"string","description":"The operation code","example":"Et sed cumque accusamus maiores."},"description":{"type":"string","description":"The operation description","example":"Omnis velit et dolores."}},"description":"Get Raw Satellite ImageResponseBody result type (default view)","example":{"code":"Officiis ut non voluptates.","description":"Quis pariatur."}},"RawImagesLoadNewRawSatelliteImageBadRequestResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Load new raw satellite image_BadRequest_Response_Body result type (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"RawImagesLoadNewRawSatelliteImageInternalErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Load new raw satellite image_InternalError_Response_Body result type (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"RawImagesLoadNewRawSatelliteImageRequestBody":{"title":"RawImagesLoadNewRawSatelliteImageRequestBody","type":"object","properties":{"file_name":{"type":"string","description":"File name of the raw image","example":"Fugiat atque voluptates sit eum rerum."},"id":{"type":"string","description":"The image identifier","example":"Fuga maxime."}},"example":{"file_name":"Laborum aliquid autem.","id":"Ullam aliquid."}},"RawImagesLoadNewRawSatelliteImageResponseBody":{"title":"Mediatype identifier: vnd.application/goa.result+json; view=default","type":"object","properties":{"code":{"type":"string","description":"The operation code","example":"Ut ut perferendis."},"description":{"type":"string","description":"The operation description","example":"Culpa harum in modi."}},"description":"Load New Raw Satellite ImageResponseBody result type (default view)","example":{"code":"Et quod sit.","description":"Dicta quo explicabo ullam molestias."}}}}