# gen/http/processed_images/server/encode_decode.go
// Code generated by goa v3.2.3, DO NOT EDIT.
//
// Processed images HTTP server encoders and decoders
//
// Command:
// $ goa gen image-loader/design

package server

import (
	"context"
	processedimages "image-loader/gen/processed_images"
	processedimagesviews "image-loader/gen/processed_images/views"
	"io"
	"net/http"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeLoadNewProcessedSatelliteImageResponse returns an encoder for
// responses returned by the Processed images Load new processed satellite
// image endpoint.
func EncodeLoadNewProcessedSatelliteImageResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(*processedimagesviews.GoaResult)
		enc := encoder(ctx, w)
		body := NewLoadNewProcessedSatelliteImageResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeLoadNewProcessedSatelliteImageRequest returns a decoder for requests
// sent to the Processed images Load new processed satellite image endpoint.
func DecodeLoadNewProcessedSatelliteImageRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body LoadNewProcessedSatelliteImageRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateLoadNewProcessedSatelliteImageRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewLoadNewProcessedSatelliteImageProcessedSatelliteImage(&body)

		return payload, nil
	}
}

// EncodeLoadNewProcessedSatelliteImageError returns an encoder for errors
// returned by the Load new processed satellite image Processed images endpoint.
func EncodeLoadNewProcessedSatelliteImageError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		en, ok := v.(ErrorNamer)
		if !ok {
			return encodeError(ctx, w, v)
		}
		switch en.ErrorName() {
		case "BadRequest":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewLoadNewProcessedSatelliteImageBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", "BadRequest")
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "InternalError":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewLoadNewProcessedSatelliteImageInternalErrorResponseBody(res)
			}
			w.Header().Set("goa-error", "InternalError")
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeGetProcessedSatelliteImageResponse returns an encoder for responses
// returned by the Processed images Get processed satellite image endpoint.
func EncodeGetProcessedSatelliteImageResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(*processedimagesviews.GoaResult)
		enc := encoder(ctx, w)
		body := NewGetProcessedSatelliteImageResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetProcessedSatelliteImageRequest returns a decoder for requests sent
// to the Processed images Get processed satellite image endpoint.
func DecodeGetProcessedSatelliteImageRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body GetProcessedSatelliteImageRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		payload := NewGetProcessedSatelliteImagePayload(&body)

		return payload, nil
	}
}

// EncodeGetProcessedSatelliteImageError returns an encoder for errors returned
// by the Get processed satellite image Processed images endpoint.
func EncodeGetProcessedSatelliteImageError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		en, ok := v.(ErrorNamer)
		if !ok {
			return encodeError(ctx, w, v)
		}
		switch en.ErrorName() {
		case "BadRequest":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewGetProcessedSatelliteImageBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", "BadRequest")
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "InternalError":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewGetProcessedSatelliteImageInternalErrorResponseBody(res)
			}
			w.Header().Set("goa-error", "InternalError")
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// unmarshalGeographicInformationRequestBodyToProcessedimagesGeographicInformation
// builds a value of type *processedimages.GeographicInformation from a value
// of type *GeographicInformationRequestBody.
func unmarshalGeographicInformationRequestBodyToProcessedimagesGeographicInformation(v *GeographicInformationRequestBody) *processedimages.GeographicInformation {
	if v == nil {
		return nil
	}
	res := &processedimages.GeographicInformation{
		TagName: v.TagName,
	}
	if v.Coordinates != nil {
		res.Coordinates = make(map[string]float64, len(v.Coordinates))
		for key, val := range v.Coordinates {
			tk := key
			tv := val
			res.Coordinates[tk] = tv
		}
	}

	return res
}

// unmarshalNormalizedIndexesRequestBodyToProcessedimagesNormalizedIndexes
// builds a value of type *processedimages.NormalizedIndexes from a value of
// type *NormalizedIndexesRequestBody.
func unmarshalNormalizedIndexesRequestBodyToProcessedimagesNormalizedIndexes(v *NormalizedIndexesRequestBody) *processedimages.NormalizedIndexes {
	if v == nil {
		return nil
	}
	res := &processedimages.NormalizedIndexes{}
	if v.Ndvi != nil {
		res.Ndvi = make([]float64, len(v.Ndvi))
		for i, val := range v.Ndvi {
			res.Ndvi[i] = val
		}
	}
	if v.Ndwi != nil {
		res.Ndwi = make([]float64, len(v.Ndwi))
		for i, val := range v.Ndwi {
			res.Ndwi[i] = val
		}
	}

	return res
}